Patch after reviewing

+++ src/core/arm/arm_interface.h
@@ -125,7 +125,7 @@ public:
      * Advance the CPU core by the specified number of ticks (e.g. to simulate CPU execution time)
      * @param ticks Number of ticks to advance the CPU core
      */
-    virtual void AddTicks(size_t ticks) = 0;
+    virtual void AddTicks(u64 ticks) = 0;
 
     /**
      * Saves the current CPU context
